+++ trip/rvl/tsopt/testsrc/test5a.cc
@@ -17,7 +17,7 @@ int main(int argc, char ** argv) {
   bool verbose=false;
   
   // define Sim object, initialize control
-  ContentPackage<float,int> c;
+  ContentPackage<float,size_t> c;
   c.initialize(n);
   for (int i=0;i<n;i++) c.getData()[i]=0.5;
   Dyn1 stp(c, dt, verbose);
@@ -34,7 +34,7 @@ int main(int argc, char ** argv) {
   StdSim<RnState> f(stp,tt, aa);
 
   // stuff for linearized step
-  ContentPackage<float,int> dc;
+  ContentPackage<float,size_t> dc;
   dc.initialize(n);
   for (int i=0;i<n;i++) dc.getData()[i]=1.0;
 
@@ -48,7 +48,7 @@ int main(int argc, char ** argv) {
   initLin<RnState> bb(dstp);
  
   // stuff for adjoint step
-  ContentPackage<float,int> ac;
+  ContentPackage<float,size_t> ac;
   ac.initialize(n);
   for (int i=0;i<n;i++) ac.getData()[i]=0.5;
