+++ trip/rvl/tsopt/testsrc/test5a_ra.cc
@@ -19,7 +19,7 @@ int main(int argc, char ** argv) {
   bool verbose=false;
   
   // define Sim object, initialize control
-  ContentPackage<float,int> c;
+  ContentPackage<float,size_t> c;
   c.initialize(n);
   for (int i=0;i<n;i++) c.getData()[i]=0.5;
   Dyn1 stp(c, dt, verbose);
@@ -36,7 +36,7 @@ int main(int argc, char ** argv) {
   RASim<RnState> f(stp,tt,aa,myStack);
 
   // stuff for linearized step
-  ContentPackage<float,int> dc;
+  ContentPackage<float,size_t> dc;
   dc.initialize(n);
   for (int i=0;i<n;i++) dc.getData()[i]=1.0;
 
@@ -51,7 +51,7 @@ int main(int argc, char ** argv) {
   initLin<RnState> bb(dstp);
 
   // stuff for adjoint step
-  ContentPackage<float,int> ac;
+  ContentPackage<float,size_t> ac;
   ac.initialize(n);
   for (int i=0;i<n;i++) ac.getData()[i]=0.5;
